pyproject.toml
[project]
name = "workspace"
run = "python3 save.py"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jinja2==3.1.4",
    "pdflatex==0.1.3",
    "pyyaml-include==2.2",
]
